EVG-14957 WAdd linter for ParserProject and the merge functions
