+++ self-tests.yml
@@ -339,6 +339,15 @@ functions:
         env:
           BRANCH_NAME: ${branch_name}
 
+  verify-merge-function-update:
+    - command: subprocess.exec
+      params:
+        working_dir: gopath/src/github.com/evergreen-ci/evergreen
+        binary: bash
+        args: ["scripts/verify-merge-function-update.sh"]
+        env:
+          BRANCH_NAME: ${branch_name}
+
   attach-test-results:
     - command: gotest.parse_files
       type: system
@@ -593,6 +602,11 @@ tasks:
     commands:
       - func: get-project
       - func: verify-agent-version-update
+  - name: verify-merge-function-update
+    patch_only: true
+    commands:
+      - func: get-project
+      - func: verify-merge-function-update
   - name: test-db-auth
     commands:
       - func: get-project
@@ -638,6 +652,7 @@ buildvariants:
       - name: ".test"
       - name: "js-test"
       - name: "verify-agent-version-update"
+      - name: "verify-merge-function-update"
       - name: "docker-cleanup"
       - name: test-db-auth
       - name: test-cloud
